WIP matejcik
WIP - more compact representation of font special characters

WIP - utf8 iterator for C

WIP - tweak get_glyph

WIP - change format of generated translations blob

WIP - drop useless generated rs files

WIP - extend trezorhal translations interface

WIP - cleaner translations interface

WIP - move translations to a top-level module

WIP - update generated files

WIP - support for make_new in obj_type!

WIP - drop modtrezorconfig implementations

WIP - functional translations module

+++ core/embed/lib/fonts/fonts.h
@@ -20,6 +20,8 @@
 #ifndef _FONTS_H
 #define _FONTS_H
 
+#include <stdbool.h>
+
 #include "fonts/font_bitmap.h"
 #include TREZOR_BOARD
 
@@ -117,7 +119,7 @@
 int font_height(int font);
 int font_max_height(int font);
 int font_baseline(int font);
-const uint8_t *font_get_glyph(int font, uint8_t c);
+const uint8_t *font_get_glyph(int font, uint16_t c);
 int font_text_width(int font, const char *text, int textlen);
 int font_text_split(int font, const char *text, int textlen,
                     int requested_width);
